+++ CPPLibrary/ContinuedFraction.cpp
@@ -2,14 +2,14 @@
 #include "ContinuedFraction.h"
 #include "Algorithms.h"
 
-__declspec(dllexport) long* CPPMathLibrary::ContinuedFractionExpansion(double gamma, size_t& count, size_t& op_count) {
+__declspec(dllexport) int* CPPMathLibrary::ContinuedFractionExpansion(double gamma, size_t& count, size_t& op_count) {
 	op_count = 0; //clear the op count
-	long* arr = new long[count]; //needs to be a signed int because q_i can be negative
-	arr[0] = (long)floor(gamma);
+	int* arr = new int[count]; //needs to be a signed int because q_i can be negative
+	arr[0] = (int)floor(gamma);
 	double gamma_n1 = 0;
 	for (size_t i = 1; i < count; ++i) {
 		gamma_n1 = 1 / (gamma - arr[i - 1]);
-		arr[i] = (long)floor(gamma_n1);
+		arr[i] = (int)floor(gamma_n1);
 		op_count += 2;
 		if (IsInt(gamma_n1) && gamma_n1 == arr[i]) {
 			count = i - 1; //terminated early so update the count with the last q
@@ -20,7 +20,7 @@ __declspec(dllexport) long* CPPMathLibrary::ContinuedFractionExpansion(double ga
 	return arr;
 }
 
-__declspec(dllexport) CPPMathLibrary::RationalNumber CPPMathLibrary::FindConvergent(long* qs, size_t stop_index) {
+__declspec(dllexport) CPPMathLibrary::RationalNumber CPPMathLibrary::FindConvergent(int* qs, size_t stop_index) {
 	if (stop_index < 0) { return RationalNumber(0, 0); /*this is treated as NaN*/ }
 	RationalNumber r(*(qs + stop_index));
 	if (stop_index == 0) { return r; }
@@ -32,7 +32,7 @@ __declspec(dllexport) CPPMathLibrary::RationalNumber CPPMathLibrary::FindConverg
 	return r;
 }
 
-__declspec(dllexport) void CPPMathLibrary::PrintConvergents(long* qs, size_t count) {
+__declspec(dllexport) void CPPMathLibrary::PrintConvergents(int* qs, size_t count) {
 	std::cout << "Convergences:" << std::endl;
 	for (size_t i = 0; i < count; ++i) {
 		std::cout << i + 1 << ") " << FindConvergent(qs, i) << std::endl;
